Fix: 修复多游戏渲染时会保留上一次场景的 Bug（每一次开始都清空资源管理器）

+++ Services/Interpreter/ScriptStructure.cs
@@ -54,7 +54,7 @@ public record struct PaintStructure
 	bool Antialias
 );
 
-public record struct ColorStructure(int R, int G, int B, int A);
+public record struct ColorStructure(byte R, byte G, byte B, byte A);
 
 public record struct ResouresStructure
 (
